chore(desktop): add React Compiler, extract useSearchPage, remove manual memoization
Enable babel-plugin-react-compiler with target "18" so memoization is
handled automatically at build time. Remove 50 manual useCallback/useMemo
wrappers across 8 files (-387 lines).

Extract useSearchPage hook from SearchPage (454 → 208 lines). Replace
userEvent with fireEvent in JobCard tests (447ms → 116ms).

Bundle size decreased ~1 KB from compiler optimizations.
