+++ crates/rdlp-desktop/package.json
@@ -28,6 +28,7 @@
     "lucide-react": "^0.575.0",
     "radix-ui": "^1.4.3",
     "react": "^18.3.1",
+    "react-compiler-runtime": "^1.0.0",
     "react-dom": "^18.3.1",
     "tailwind-merge": "^3.5.0",
     "tailwindcss": "^4.2.0",
@@ -42,6 +43,7 @@
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.0",
     "@vitest/coverage-v8": "^4.0.18",
+    "babel-plugin-react-compiler": "^1.0.0",
     "cypress": "^15.10.0",
     "happy-dom": "^20.8.3",
     "jsdom": "^28.1.0",
