+++ crates/rdlp-desktop/src/components/FormatDialog.tsx
@@ -1,6 +1,6 @@
 // Full-screen modal for format selection with hero table, presets, and collapsed options.
 
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import { useQuery } from "@tanstack/react-query";
 import {
     useReactTable,
@@ -118,14 +118,11 @@ export function FormatDialog({ url, onConfirm, onClose }: FormatDialogProps) {
 
     const groups = groupRows(table.getSortedRowModel().rows);
 
-    const subtitleLangs = useMemo(
-        () => formatData?.subtitles.map((s) => s.lang) ?? [],
-        [formatData],
-    );
+    const subtitleLangs = formatData?.subtitles.map((s) => s.lang) ?? [];
 
     // -- Handlers -----------------------------------------------------
 
-    const handleRowClick = useCallback((id: string, e: React.MouseEvent) => {
+    const handleRowClick = (id: string, e: React.MouseEvent) => {
         if (e.shiftKey && selectedId !== null && selectedId !== id) {
             setMergeId(id);
         } else {
@@ -135,9 +132,9 @@ export function FormatDialog({ url, onConfirm, onClose }: FormatDialogProps) {
         // Clear preset when manually selecting
         setPresetId(null);
         setOptions((prev) => ({ ...prev, format: null }));
-    }, [selectedId]);
+    };
 
-    const handlePresetSelect = useCallback((newPresetId: string) => {
+    const handlePresetSelect = (newPresetId: string) => {
         if (!newPresetId) {
             setPresetId(null);
             setOptions((prev) => ({ ...prev, format: null }));
@@ -149,14 +146,14 @@ export function FormatDialog({ url, onConfirm, onClose }: FormatDialogProps) {
         setSelectedId(null);
         setMergeId(null);
         setOptions((prev) => ({ ...prev, format: preset.selector }));
-    }, []);
+    };
 
-    const handleClearSelection = useCallback(() => {
+    const handleClearSelection = () => {
         setSelectedId(null);
         setMergeId(null);
-    }, []);
+    };
 
-    const handleExprChange = useCallback((value: string) => {
+    const handleExprChange = (value: string) => {
         setExpr(value);
         if (exprTimerRef.current) clearTimeout(exprTimerRef.current);
         if (value.trim() === "") {
@@ -188,26 +185,26 @@ export function FormatDialog({ url, onConfirm, onClose }: FormatDialogProps) {
                 setExprMatches(new Set());
             }
         }, 400);
-    }, [formatData]);
+    };
 
-    const buildFormatSelector = useCallback((): string | null => {
+    const buildFormatSelector = (): string | null => {
         if (expertMode && expr.trim()) return expr.trim();
         if (selectedId && mergeId) return `${selectedId}+${mergeId}`;
         if (selectedId) return selectedId;
         return options.format;
-    }, [expertMode, expr, selectedId, mergeId, options.format]);
+    };
 
-    const handleConfirm = useCallback(() => {
+    const handleConfirm = () => {
         const format = buildFormatSelector();
         onConfirm({ ...options, format }, formatData?.title);
-    }, [buildFormatSelector, onConfirm, options, formatData?.title]);
+    };
 
-    const handleBrowseDir = useCallback(async () => {
+    const handleBrowseDir = async () => {
         const dir = await pickDirectory();
         if (dir) setOptions((prev) => ({ ...prev, outputDir: dir }));
-    }, []);
+    };
 
-    const handleSubLangSelect = useCallback((lang: string) => {
+    const handleSubLangSelect = (lang: string) => {
         setOptions((prev) => {
             const current = prev.subtitleLangs;
             const next = current.includes(lang)
@@ -215,14 +212,11 @@ export function FormatDialog({ url, onConfirm, onClose }: FormatDialogProps) {
                 : [...current, lang];
             return { ...prev, subtitleLangs: next };
         });
-    }, []);
+    };
 
     // -- Lookup helpers -----------------------------------------------
 
-    const findFormat = useCallback(
-        (id: string) => formatData?.formats.find((f) => f.format_id === id),
-        [formatData],
-    );
+    const findFormat = (id: string) => formatData?.formats.find((f) => f.format_id === id);
 
     // -- Render -------------------------------------------------------
